exclude app from main for different entries
