+++ lib/app_export.dart
@@ -1,2 +1,3 @@
+export 'my_app.dart';
 export 'service_locator.dart';
 export 'skeleton/export.dart';
